# optimized parse_colmap_output by 99x