The optimized version is significantly faster because it replaces inefficient per-element Python loops and repeated small tensor operations with batched PyTorch operations.

Benchmarking with 100 images, 2 cameras on cuda
Validation passed
Original time: 0.0821s
Optimized time: 0.0025s
Speedup: 33.04x

Benchmarking with 1000 images, 5 cameras on cuda
Validation passed
Original time: 0.6518s
Optimized time: 0.0069s
Speedup: 94.17x

Benchmarking with 10000 images, 10 cameras on cuda
Validation passed
Original time: 6.5895s
Optimized time: 0.0778s
Speedup: 84.72x


here's the benchmarking and validation script
https://colab.research.google.com/drive/1dBQY9i_pYwMFICIJt_VJTGs9QmlBlJa-?usp=sharing